+++ error.go
@@ -8,6 +8,14 @@ func NoError(err error) FailureMessage {
 	return "non-nil error:\n" + FailureMessage(err.Error())
 }
 
+// IsError tests if the error is non-nil.
+func IsError(err error) FailureMessage {
+	if err != nil {
+		return ""
+	}
+	return "the error is <nil>"
+}
+
 // FluentError encapsulates assertions for error object.
 type FluentError struct {
 	FluentAny[error]